Moved most of TTS-related code out of AbstractFlashcardViewer.
