+++ AnkiDroid/src/main/java/com/ichi2/anki/AbstractFlashcardViewer.java
@@ -39,7 +39,6 @@ import android.os.Handler;
 import android.os.Message;
 import android.os.SystemClock;
 import android.os.Vibrator;
-import android.speech.tts.TextToSpeech;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.view.GestureDetectorCompat;
 import android.support.v7.app.ActionBar;
@@ -99,19 +98,14 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.lang.ref.WeakReference;
 import java.net.URLDecoder;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import timber.log.Timber;
-
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Element;
 
 public abstract class AbstractFlashcardViewer extends NavigationDrawerActivity {
 
@@ -2261,48 +2255,13 @@ public abstract class AbstractFlashcardViewer extends NavigationDrawerActivity {
     }
 
     /**
-     * Snippet of text accompanied by its locale code (if known).
-     */
-    private static final class LocalisedText {
-        private String mText;
-        private String mLocaleCode;
-
-        /**
-         * Construct an object representing a snippet of text in an unknown locale.
-         */
-        public LocalisedText(String text) {
-            mText = text;
-            mLocaleCode = "";
-        }
-
-        /**
-         * Construct an object representing a snippet of text in a particular locale.
-         *
-         * @param localeCode A string representation of a locale in the format returned by
-         *                    Locale.toString().
-         */
-        public LocalisedText(String text, String localeCode) {
-            mText = text;
-            mLocaleCode = localeCode;
-        }
-
-        public String getText() {
-            return mText;
-        }
-
-        public String getLocaleCode() {
-            return mLocaleCode;
-        }
-    }
-
-    /**
      * Reads the text (using TTS) for the given side of a card.
      *
      * @param card     The card to play TTS for
      * @param cardSide The side of the current card to play TTS for
      */
     private static void readCardText(final Card card, final int cardSide) {
-        String cardSideContent;
+        final String cardSideContent;
         if (Sound.SOUNDS_QUESTION == cardSide) {
             cardSideContent = card.q(true);
         } else if (Sound.SOUNDS_ANSWER == cardSide) {
@@ -2312,53 +2271,7 @@ public abstract class AbstractFlashcardViewer extends NavigationDrawerActivity {
             return;
         }
 
-        long deckId = getDeckIdForCard(card);
-        int ord = card.getOrd();
-        boolean isFirstText = true;
-        for (LocalisedText textToRead : getTextsToRead(cardSideContent)) {
-            if (!textToRead.getText().isEmpty()) {
-                ReadText.textToSpeech(textToRead.getText(), deckId, ord, cardSide,
-                        textToRead.getLocaleCode(),
-                        isFirstText ? TextToSpeech.QUEUE_FLUSH : TextToSpeech.QUEUE_ADD);
-                isFirstText = false;
-            }
-        }
-    }
-
-    /**
-     * Returns the list of text snippets contained in the given HTML fragment that should be read
-     * using the Android text-to-speech engine, together with the languages they are in.
-     * <p>
-     * Each returned LocalisedText object contains the text extracted from a &lt;tts&gt; element
-     * whose 'service' attribute is set to 'android', and the localeCode taken from the 'voice'
-     * attribute of that element. This holds unless the HTML fragment contains no such &lt;tts&gt;
-     * elements; in that case the function returns a single LocalisedText object containing the
-     * text extracted from the whole HTML fragment, with the localeCode set to an empty string.
-     */
-    private static List<LocalisedText> getTextsToRead(String html) {
-        List<LocalisedText> textsToRead = new ArrayList<>();
-
-        Element elem = Jsoup.parseBodyFragment(html).body();
-        parseTtsElements(elem, textsToRead);
-        if (textsToRead.size() == 0) {
-            // No <tts service="android"> elements found: return the text of the whole HTML fragment
-            textsToRead.add(new LocalisedText(elem.text()));
-        }
-
-        return textsToRead;
-    }
-
-    private static void parseTtsElements(Element element, List<LocalisedText> textsToRead) {
-        if (element.tagName().equalsIgnoreCase("tts")) {
-            if (element.attr("service").equalsIgnoreCase("android")) {
-                textsToRead.add(new LocalisedText(element.text(), element.attr("voice")));
-                return; // ignore any children
-            }
-        }
-
-        for (Element child : element.children()) {
-            parseTtsElements(child, textsToRead);
-        }
+        ReadText.readCardSide(cardSide, cardSideContent, getDeckIdForCard(card), card.getOrd());
     }
 
     /**
